Remove scan-prs action
