+++ README.md
@@ -146,7 +146,7 @@ Free up disk space on GitHub runners by removing unnecessary packages and files.
 
 ### 3. GitHub Report Action
 
-Summarize open PRs and failed default-branch GitHub Actions across organization repositories.
+Summarize failed default-branch GitHub Actions across organization repositories.
 
 ```yaml
 - uses: ultralytics/actions/github-report@main
@@ -185,7 +185,6 @@ pip install ultralytics-actions
 - `actions.review_pr` - AI-powered PR review
 - `actions.summarize_pr` - Generate PR summaries
 - `actions.github_report` - Generate GitHub organization reports
-- `actions.scan_prs` - Scan and manage organization PRs
 - `actions.first_interaction` - Welcome message for new contributors
 - And more in `actions/` directory
