+++ actions/__init__.py
@@ -16,7 +16,6 @@
 # │   ├── dispatch_actions.py
 # │   ├── first_interaction.py
 # │   ├── review_pr.py
-# │   ├── scan_prs.py
 # │   ├── dependabot.py
 # │   ├── summarize_pr.py
 # │   ├── summarize_release.py
